+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Migrate out-of-band remote access decoder to stream API
 - Migrate system configuration options decoder to stream API
 - Migrate system reset information decoder to stream API
 - Migrate system information decoder to stream API